requirements.txt
google-cloud-bigquery==3.17.2
pandas==2.2.1
python-dotenv==1.0.1
tabulate==0.9.0

